QuadratureGenerator with discrete levelset
