+++ tests/non_matching/discrete_quadrature_generator.output
@@ -1,0 +1,40 @@
+
+DEAL::dim = 1
+DEAL::0.375000, 0.750000
+DEAL::0.875000, 0.250000
+DEAL::0.750000, 1.00000, 1.00000
+DEAL::1.00000
+DEAL::1.00000
+DEAL::
+DEAL::dim = 2
+DEAL::0.375000, 0.500000, 0.750000
+DEAL::0.875000, 0.500000, 0.250000
+DEAL::0.750000, 0.500000, 1.00000, 1.00000, 0.00000
+DEAL::0.500000, 1.00000
+DEAL::0.500000, 1.00000
+DEAL::0.375000, 0.750000
+DEAL::0.875000, 0.250000
+DEAL::0.750000, 1.00000, 1.00000, 0.00000
+DEAL::0.375000, 0.750000
+DEAL::0.875000, 0.250000
+DEAL::0.750000, 1.00000, 1.00000, 0.00000
+DEAL::
+DEAL::dim = 3
+DEAL::0.375000, 0.500000, 0.500000, 0.750000
+DEAL::0.875000, 0.500000, 0.500000, 0.250000
+DEAL::0.750000, 0.500000, 0.500000, 1.00000, 1.00000, 0.00000, 0.00000
+DEAL::0.500000, 0.500000, 1.00000
+DEAL::0.500000, 0.500000, 1.00000
+DEAL::0.500000, 0.375000, 0.750000
+DEAL::0.500000, 0.875000, 0.250000
+DEAL::0.500000, 0.750000, 1.00000, 1.00000, 0.00000, 0.00000
+DEAL::0.500000, 0.375000, 0.750000
+DEAL::0.500000, 0.875000, 0.250000
+DEAL::0.500000, 0.750000, 1.00000, 1.00000, 0.00000, 0.00000
+DEAL::0.375000, 0.500000, 0.750000
+DEAL::0.875000, 0.500000, 0.250000
+DEAL::0.750000, 0.500000, 1.00000, 1.00000, 0.00000, 0.00000
+DEAL::0.375000, 0.500000, 0.750000
+DEAL::0.875000, 0.500000, 0.250000
+DEAL::0.750000, 0.500000, 1.00000, 1.00000, 0.00000, 0.00000
+DEAL::
